// projet-stagiaire1/src/app/model/encadrant.ts
// src/app/model/encadrant.ts
export interface Encadrant {
  idpersonne: number;
  nom: string;
  prenom: string;
  datenaiss: string;
  lieunaiss: string;
  login: string;
  motpasse: string;
  telephonne: string;
  email: string;
  matricule: string;
  cin: string;
  etats: string;
  specialite: string;
}